refactor: created pagination helper class to encapsulate duplicated code in functions
