+++ src/Shared/Common/Common.csproj
@@ -7,6 +7,8 @@
 
   <ItemGroup>
     <PackageReference Include="Azure.Storage.Blobs" Version="12.21.2" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.13" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.19.0" />
   </ItemGroup>
 
   <PropertyGroup>
